chore: move some folders to internal
