+++ .gitignore
@@ -4,7 +4,6 @@
 *.dll
 *.so
 *.dylib
-fga
 
 # Test binary, built with `go test -c`
 *.test
@@ -20,3 +19,6 @@ go.work
 
 # IDEs
 .idea/
+
+# Built files
+dist/
