+++ cmd/model/get_test.go
@@ -7,8 +7,8 @@ import (
 	"testing"
 
 	"github.com/golang/mock/gomock"
-	"github.com/openfga/cli/lib/fga"
-	mock_client "github.com/openfga/cli/mocks"
+	"github.com/openfga/cli/internal/fga"
+	"github.com/openfga/cli/internal/mocks"
 	openfga "github.com/openfga/go-sdk"
 	"github.com/openfga/go-sdk/client"
 )
