+++ cmd/query/expand_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/golang/mock/gomock"
-	mock_client "github.com/openfga/cli/mocks"
+	"github.com/openfga/cli/internal/mocks"
 	"github.com/openfga/go-sdk/client"
 )
