+++ cmd/query/list-objects.go
@@ -19,8 +19,8 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/openfga/cli/lib/cmd-utils"
-	"github.com/openfga/cli/lib/output"
+	cmdutils2 "github.com/openfga/cli/internal/cmdutils"
+	"github.com/openfga/cli/internal/output"
 	"github.com/openfga/go-sdk/client"
 	"github.com/spf13/cobra"
 )
@@ -57,14 +57,14 @@ var listObjectsCmd = &cobra.Command{
 	Args:    cobra.ExactArgs(3),                                                                                                                                                                                            //nolint:gomnd,lll
 	Example: `fga query list-objects --store-id=01H0H015178Y2V4CX10C2KGHF4 user:anne can_view document --contextual-tuple "user:anne can_view folder:product" --contextual-tuple "folder:product parent document:roadmap"`, //nolint:lll
 	RunE: func(cmd *cobra.Command, args []string) error {
-		clientConfig := cmdutils.GetClientConfig(cmd)
+		clientConfig := cmdutils2.GetClientConfig(cmd)
 
 		fgaClient, err := clientConfig.GetFgaClient()
 		if err != nil {
 			return fmt.Errorf("failed to initialize FGA Client due to %w", err)
 		}
 
-		contextualTuples, err := cmdutils.ParseContextualTuples(cmd)
+		contextualTuples, err := cmdutils2.ParseContextualTuples(cmd)
 		if err != nil {
 			return fmt.Errorf("error parsing contextual tuples for listObjects: %w", err)
 		}
