+++ cmd/query/list-relations_test.go
@@ -8,8 +8,8 @@ import (
 	"testing"
 
 	"github.com/golang/mock/gomock"
-	"github.com/openfga/cli/lib/fga"
-	mock_client "github.com/openfga/cli/mocks"
+	"github.com/openfga/cli/internal/fga"
+	"github.com/openfga/cli/internal/mocks"
 	"github.com/openfga/go-sdk/client"
 )
