+++ cmd/tuple/delete.go
@@ -19,8 +19,8 @@ import (
 	"context"
 	"fmt"
 
-	cmdutils "github.com/openfga/cli/lib/cmd-utils"
-	"github.com/openfga/cli/lib/output"
+	"github.com/openfga/cli/internal/cmdutils"
+	"github.com/openfga/cli/internal/output"
 	"github.com/openfga/go-sdk/client"
 	"github.com/spf13/cobra"
 )
