+++ internal/output/marshal.go
